test: Add pytest and restructure examples
